+++ backend/agenda_upload_service/lambda/routes/index.js
@@ -1,0 +1,11 @@
+const { Router } = require('express');
+const getUploadController = require('../controllers/uploadController');
+
+module.exports = (logger) => {
+  const router = Router();
+  const controller = getUploadController(logger);
+
+  router.post('/upload', controller.csvUpload);
+
+  return router;
+};
